requirements.txt
python-dotenv
jokeapi
discord.py[voice]
spotipy
yt_dlp
pylast
ffmpeg
groq
validators
soundcloud-lib